+++ src/main/java/christmas/model/Badge.java
@@ -13,14 +13,14 @@ public enum Badge {
     SANTA("산타", 20_000);
 
     private final String name;
-    private final int benefitAmount;
+    private final long benefitAmount;
 
-    Badge(String name, int benefitAmount) {
+    Badge(String name, long benefitAmount) {
         this.name = name;
         this.benefitAmount = benefitAmount;
     }
 
-    public static String getBadgeNameOf(int benefitAmount) {
+    public static String getBadgeNameOf(long benefitAmount) {
         String resultBadgeName = null;
         for(Badge badge : Badge.values()){
             if(badge.benefitAmount <= benefitAmount)
